blender_actions.py is now a blender_actions module, and the monolith file is split up into several files.
