+++ providers/blender/blender_provider/blender_actions/__init__.py
@@ -1,0 +1,22 @@
+# Actions wrap Blender's API to perform a single action.
+# An implementation of an action should avoid performing any logic
+# An implementation of an action is allowed to perform unit conversions or perform read operations for pre-checks.
+
+from .objects import *
+from .collections import *
+from .addons import *
+from .context import *
+from .nodes import *
+from .scene import *
+from .animation import *
+from .camera import *
+from .drivers import *
+from .curve import *
+from .transformations import *
+from .constraints import *
+from .import_export import *
+from .light import *
+from .material import *
+from .mesh import *
+from .modifiers import *
+from .render import *
